requirements.txt
openai>=1.57.0
python-dotenv==1.0.0
json5==0.9.14